+++ languages/de.json
@@ -1,5 +1,6 @@
 {
     "TAH_WITCHER": {
-        "combat": "Kampf"
+        "combat": "Kampf",
+        "Weapons": "Waffen"
     }
 }
