+++ scripts/constants.js
@@ -1,0 +1,7 @@
+// IDs
+export const COMBAT_ID = 'Combat';
+export const WEAPONS_ID = 'Weapons';
+
+export const ACTIONS = {
+    ATTACK: 'action_attack'
+};
